Renamed transaction-related DTO classes to remove the "Response" suffix and updated references across the data, model, and network layers.
