+++ core/network/src/commonMain/kotlin/org/mifos/mobile/core/network/services/LoanAccountsListService.kt
@@ -18,7 +18,7 @@ import de.jensklingenberg.ktorfit.http.Query
 import io.ktor.client.statement.HttpResponse
 import kotlinx.coroutines.flow.Flow
 import org.mifos.mobile.core.model.entity.accounts.loan.LoanAccount
-import org.mifos.mobile.core.model.entity.accounts.loan.LoanTransactionDetailsResponse
+import org.mifos.mobile.core.model.entity.accounts.loan.LoanTransactionDetails
 import org.mifos.mobile.core.model.entity.accounts.loan.LoanWithAssociations
 import org.mifos.mobile.core.model.entity.accounts.loan.LoanWithdraw
 import org.mifos.mobile.core.model.entity.payload.LoansPayload
@@ -63,5 +63,5 @@ interface LoanAccountsListService {
     fun getLoanTransactionDetails(
         @Path("loanId") loanId: Long,
         @Path("transactionId") transactionId: Long,
-    ): Flow<LoanTransactionDetailsResponse>
+    ): Flow<LoanTransactionDetails>
 }
